fix location of transformation files
