+++ litellm/__init__.py
@@ -855,16 +855,28 @@ from .llms.bedrock.chat.invoke_handler import (
 from .llms.bedrock.common_utils import (
     AmazonBedrockGlobalConfig,
 )
-from .llms.bedrock.chat.invoke_transformations.bedrock_transformations import (
-    AmazonTitanConfig,
+from .llms.bedrock.chat.invoke_transformations.amazon_ai21_transformation import (
     AmazonAI21Config,
+)
+from .llms.bedrock.chat.invoke_transformations.amazon_anthropic_claude2_transformation import (
     AmazonAnthropicConfig,
+)
+from .llms.bedrock.chat.invoke_transformations.amazon_anthropic_claude3_transformation import (
     AmazonAnthropicClaude3Config,
+)
+from .llms.bedrock.chat.invoke_transformations.amazon_cohere_transformation import (
     AmazonCohereConfig,
+)
+from .llms.bedrock.chat.invoke_transformations.amazon_llama_transformation import (
     AmazonLlamaConfig,
+)
+from .llms.bedrock.chat.invoke_transformations.amazon_mistral_transformation import (
     AmazonMistralConfig,
 )
-from .llms.bedrock.chat.invoke_transformations.bedrock_transformations import (
+from .llms.bedrock.chat.invoke_transformations.amazon_titan_transformation import (
+    AmazonTitanConfig,
+)
+from .llms.bedrock.chat.invoke_transformations.base_invoke_transformation import (
     AmazonInvokeConfig,
 )
